+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 0.17.1
+
 ## 0.17.0
 
 * Log all request failures by default rather than just errors after the first one, can now be set with the `retry_log_threshold` argument to `configure()`.  To get the previous behavior, use `bf.configure(retry_log_threshold=1)`